pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "water-system-simulator-sdk"
version = "0.1.0"
authors = [
  { name="Jules", email="jules@example.com" },
]
description = "A Python SDK for simulating water systems."
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "pyyaml",
    "pandas",
    "numpy",
    "matplotlib",
    "scipy", # MPC controller likely needs this
]

[project.urls]
"Homepage" = "https://github.com/example/water-system-simulator"
"Bug Tracker" = "https://github.com/example/water-system-simulator/issues"

[tool.setuptools.packages.find]
where = ["src"]
